bitkeeper revision 1.1159.1.289 (417fb030yj7vBz8Pvd4b0uvFxr0PzQ)
Ensure we use 64-bit arithmetic when converting a possibly-large
microseconds value into a nanoseconds value (e.g., when calculating
current system time in the timer interrupt handler).

+++ linux-2.4.27-xen-sparse/arch/xen/kernel/time.c
@@ -397,7 +397,7 @@ static inline void do_timer_interrupt(int irq, void *dev_id,
         __get_time_values_from_xen();
         
         delta = (s64)(shadow_system_time + 
-                      (__get_time_delta_usecs() * 1000) -
+                      ((s64)__get_time_delta_usecs() * 1000LL) -
                       processed_system_time);
     }
     while ( !TIME_VALUES_UP_TO_DATE );
